### 2021-02-04 #3003 @whedon
[PRE REVIEW]: kuibit: Analyzing Einstein Toolkit simulations with Python
**Submitting author:** @sbozzolo (<a href="http://orcid.org/0000-0003-3696-6408">Gabriele Bozzola</a>)
**Repository:** <a href="https://github.com/Sbozzolo/kuibit/" target ="_blank">https://github.com/Sbozzolo/kuibit/</a>
**Version:** 1.0.0
