### 2021-02-04 #3003 @Sbozzolo
I never entered the branch where the paper is (did I miss that?), so I guess that's the problem. The paper is in the `joss-paper` branch and compiles successfully.
