+++ paper.bib
@@ -20,6 +20,14 @@
   url = {https://simulationtools.org}
 }
 
+@misc{pycactus,
+  author = {Kastaun, Wolfgang},
+  title = {PyCactus},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/wokast/PyCactus}
+}
+
 @InProceedings{cactuscode,
   suggested-for ={Cactus/Cactus},
   author =       {Tom Goodale and Gabrielle Allen and Gerd Lanfermann
