### 2021-02-04 #3003 @eloisabentivegna
> Hi @eloisabentivegna, the author suggested you to edit this submission—I also see that you contributed to the Einstein Toolkit itself, so this seems like a great fit!

Indeed @kyleniemeyer, this is right up my alley. 🙂
