### 2021-02-12 7ba7df7
Add docstring to __getitem__ in HierarchicalGridData
