### 2021-02-12 782d3e3
Expand grid_data documentation
