+++ docs/grid_data.rst
@@ -82,17 +82,17 @@ The operation considers the size of the cell, for example
 The :py:meth:`~.contains` is syntactic sugar for the same operation.
 
 To obtain all the coordinates in the grid, you can use the
-:py:meth:`~.grid_data.UnfiromGrid.coordinates` method. This can be used in thre different ways. When
-called with no arguments, the output is a list of 1D arrays. Each of these
-arrays contains the coordinates along a fixed axis. For example, for the 2D
-grid, the first array will be the x coordinates, the second the y. Finally, with
-``as_meshgrid=True``, the return value will be a NumPy meshgrid. This is useful
-for plotting. When ``as_same_shape=True`` the return value is a list of
-coordinates with the same shape of the grid itself, each element of this list is
-the value of that coordinate over the grid. This last one is the most useful way
-to do computations that involve the coordinates.
-You can obtained the coordinate as a list of coordinates along each direction
-also with the method :py:meth:`~.coordinates_1d`.
+:py:meth:`~.grid_data.UnfiromGrid.coordinates` method. This can be used in three
+different ways. When called with no arguments, the output is a list of 1D
+arrays. Each of these arrays contains the coordinates along a fixed axis. For
+example, for the 2D grid, the first array will be the x coordinates, the second
+the y. Finally, with ``as_meshgrid=True``, the return value will be a NumPy
+meshgrid. This is useful for plotting. When ``as_same_shape=True`` the return
+value is a list of coordinates with the same shape of the grid itself, each
+element of this list is the value of that coordinate over the grid. This last
+one is the most useful way to do computations that involve the coordinates. You
+can obtained the coordinate as a list of coordinates along each direction also
+with the method :py:meth:`~.coordinates_1d`.
 
 To obtain a coordinate from a multidimensional index, just use the bracket
 operator (``box[i, j]``).
@@ -188,10 +188,10 @@ Some basic useful functions are :py:meth:`~.mean`, :py:meth:`~.integral`,
 with :math:`\Delta v` being the volume of a cell.
 
 :py:class:`~.UniformGridData` can be derived along a direction with
-:py:meth:`~.grid_data.UnfiromGridData.partial_derived`, or the gradient can be calculated with meth:`~.grid_data.UnfiromGridData.gradient`.
-In both cases, the order of the derivative can be specified. The derivative
-are numerical with finite difference. Derivative are second order accurate
-everywhere.
+:py:meth:`~.grid_data.UnfiromGridData.partial_derived`, or the gradient can be
+:py:calculated with meth:`~.grid_data.UnfiromGridData.gradient`. In both cases,
+:py:the order of the derivative can be specified. The derivative are numerical
+:py:with finite difference. Derivative are second order accurate everywhere.
 
 A convenient function is :py:meth:`~.sample_function`. This takes a multivariate
 function (e.g., :math:`sin(x + y)`) and returns a :py:class:`~.UniformGridData`
@@ -205,14 +205,15 @@ function can either be relative (percentuals, as 0.01, 0.5, or so, if you enable
 ``relative=True``), or the actual number of points.
 
 You can resample the data to a new grid using the function
-:py:meth:`~.grid_data.UniformGridData.resampled`, which takes as input a :py:class:`~.UniformGrid` and
-returns a new :py:class:`~.UniformGridData` resampled on the new grid. If the
-new grid is outside the old one, you can either raise an error, of fill the
-points outside with zeros. This behavior is controlled by the flag ``ext``. When
-``ext=1``, zeros are returned, when it is 2, ``ValueError`` is raised. By
-default, :py:meth:`~.grid_data.UniformGridData.resampled` uses a multilinear interpolation, but you can
-force to use a piecewise constant interpolation with the nearest neighbors by
[…]
