### 2021-02-14 d8dc3b0
Merge if statemements
