+++ kuibit/series.py
@@ -468,9 +468,8 @@ class BaseSeries(BaseNumerical):
 
         """
         # If x is the same, there's no need to resample
-        if len(self.x) == len(new_x):
-            if np.allclose(self.x, new_x, atol=1e-14):
-                return self.copy()
+        if len(self.x) == len(new_x) and np.allclose(self.x, new_x, atol=1e-14):
+            return self.copy()
 
         # Unfortunately there is no nearest neighor resampling in SciPy's splines.
         # Hence, we use directly the method interp1d.
