### 2021-02-14 b6107bb
Raise NotImplementedError when calling abstract methods
