### 2021-02-15 345f9e0
Fix rtype in iter_from_finest
