### 2021-02-15 82f4230
Add nanmax, nanmin, abs_nanmax, abs_nanmin
