+++ _sources/examples/grid_data.ipynb.txt
@@ -160,7 +160,8 @@
     "array. Often, we want to plot these objects. Using `data` directly would lead to \n",
     "transposing the actual physical quantity, because `data` is a matrix stored by rows, \n",
     "so, the first index does not label the x coordiante, but the y. \n",
-    "We provide the aliaxs `data_xyz`, which is the the tranposed of `data`. This is ready to be plotted. \n"
+    "We provide the aliaxs `data_xyz`, which is the the tranposed of `data`. This is ready to be plotted. \n",
+    "Here we plot the various quantities directly so we can see how to access the data. `kuibit.visualize` contains methods to streamline these operations.\n"
    ]
   },
   {
