### 2021-02-15 80a45b8
Remove duplicated iterations in multipoles from hdf5
