### 2021-02-15 ecfe36e
Rename remove_duplicate_iters to rename_duplicated_iters
