### 2021-02-15 09f3985
Add NEWS.md
