+++ NEWS.md
@@ -1,0 +1,12 @@
+# Changelog
+
+##  Version 1.0.0b1 (currently in development)
+
+#### General
+- Improvements to documentation and docstrings
+#### Features
+- Added `nanmax`, `nanmix`, `abs_nanmax`, `abs_nanmin` to `BaseNumerical`
+#### Bug fixes
+- `cactus_horizons` and `cactus_multiploes` now remove duplicate iterations
+#### Breaking changes
+- `remove_duplicate_iters` was renamed to `remove_duplicated_iters`
