+++ _sources/series.rst.txt
@@ -488,7 +488,7 @@ Common operations like time-shifting a series so that the absolute maximum (or
 minimum) is at ``t=0`` have specialized methods (for convenience):
 ``align_at_maximum`` and ``align_at_minimum``.
 
-remove_duplicate_iters
+remove_duplicated_iters
 ^^^^^^^^^^^^^^^^^^^^^^
 
 This function takes two arrays ``t`` and ``y`` and remove overlapping segments
