+++ _sources/index.rst.txt
@@ -43,11 +43,17 @@ For a full list of available features, see the :doc:`features page <features>`.
   ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Handle unit conversion, in particular from geometrized to physical
   (:py:mod:`~.unitconv`).
+- [EXPERIMENTAL] Visualize data with ``matplotlib``
+  (:py:mod:`~.visualize_matplotlib`).
+- [EXPERIMENTAL] Write command-line scripts (:py:mod:`~.argparse_helper`).
+- [EXPERIMENTAL] Make movies with  `motionpicture`_.
+
+.. _motionpicture: https://github.com/Sbozzolo/motionpicture
 
 Installation
 ------------
 
-``kuibit`` is available in PyPI. To install it with `pip`
+``kuibit`` is available in PyPI. To install it with ``pip``
 
 .. code-block:: bash
 
@@ -85,6 +91,8 @@ Usage
    grid_data.rst
    sensitivity_curves.rst
    unitconv.rst
+   visualize_matplotlib.rst
+   argparse_helper.rst
 
 Examples
 --------
@@ -98,6 +106,8 @@ Examples
    examples/cactus_grid_functions.ipynb
    examples/cactus_horizons.ipynb
    examples/gravitational_waves.ipynb
+
+You can additional examples in the folder ``examples`` in the ``experimental``.
 
 Reference material (classes, functions, ...)
 ---------------------------------------------
@@ -120,6 +130,8 @@ Reference material (classes, functions, ...)
    grid_data_ref.rst
    sensitivity_curves_ref.rst
    unitconv_ref.rst
+   visualize_matplotlib_ref.rst
+   argparse_helper_ref.rst
 
 What is a kuibit?
 -----------------
