+++ _sources/examples/cactus_grid_functions.ipynb.txt
@@ -153,8 +153,7 @@
    "metadata": {},
    "source": [
     "The resolution is poor because we are only using 10 points (since this documentation is generated on the fly, so we want to keep the computation cost small). \n",
-       "\n",
-       "Here we are plotting the quantities directly so we can how how to access the data. However, `kuibit` has also a module for plotting, which simplifies most of these operations.\n",
+    "\n",
     "We can also make 3D plots. A simple package to made 3D plots is `mayavi`."
    ]
   },
