+++ _sources/index.rst.txt
@@ -43,17 +43,11 @@ For a full list of available features, see the :doc:`features page <features>`.
   ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Handle unit conversion, in particular from geometrized to physical
   (:py:mod:`~.unitconv`).
-- [EXPERIMENTAL] Visualize data with ``matplotlib``
-  (:py:mod:`~.visualize_matplotlib`).
-- [EXPERIMENTAL] Write command-line scripts (:py:mod:`~.argparse_helper`).
-- [EXPERIMENTAL] Make movies with  `motionpicture`_.
-
-.. _motionpicture: https://github.com/Sbozzolo/motionpicture
 
 Installation
 ------------
 
-``kuibit`` is available in PyPI. To install it with ``pip``
+``kuibit`` is available in PyPI. To install it with `pip`
 
 .. code-block:: bash
 
@@ -74,6 +68,9 @@ Users and developers of ``kuibit`` meet in the `Telegram group
 <https://t.me/kuibit>`_. If you have any problem or suggestion, that's a good
 place where to discuss it. Alternatively, you can also open an issue on GitHub.
 
+Frequently asked questions are collected in the page :ref:`faq:Frequently Asked
+Questions`.
+
 Usage
 -----
 
@@ -91,8 +88,6 @@ Usage
    grid_data.rst
    sensitivity_curves.rst
    unitconv.rst
-   visualize_matplotlib.rst
-   argparse_helper.rst
 
 Examples
 --------
@@ -107,8 +102,6 @@ Examples
    examples/cactus_horizons.ipynb
    examples/gravitational_waves.ipynb
 
-You can additional examples in the folder ``examples`` in the ``experimental``.
-
 Reference material (classes, functions, ...)
 ---------------------------------------------
 
@@ -116,6 +109,7 @@ Reference material (classes, functions, ...)
    :maxdepth: 1
 
    features.rst
+   faq.rst
    simdir_ref.rst
    series_ref.rst
    timeseries_ref.rst
@@ -130,8 +124,6 @@ Reference material (classes, functions, ...)
    grid_data_ref.rst
    sensitivity_curves_ref.rst
    unitconv_ref.rst
-   visualize_matplotlib_ref.rst
-   argparse_helper_ref.rst
 
 What is a kuibit?
 -----------------
