+++ _sources/faq.rst.txt
@@ -15,6 +15,7 @@ reduced.
 
 To solve this problem, you have to reduce the number of files in your directory.
 There are two ways to do this:
+
 1. you create a new folder filled with files obtained by concatenating the
    corresponding files across different simulations restarts. For example, if
    your simulation has 10 restarts, each one with the same file `rho.x.asc`, you
