+++ docs/index.rst
@@ -89,18 +89,18 @@ Usage
    sensitivity_curves.rst
    unitconv.rst
 
-Examples
---------
+Tutorials
+---------
 
 .. toctree::
    :maxdepth: 1
 
-   examples/simdir.ipynb
-   examples/timeseries.ipynb
-   examples/grid_data.ipynb
-   examples/cactus_grid_functions.ipynb
-   examples/cactus_horizons.ipynb
-   examples/gravitational_waves.ipynb
+   tutorials/simdir.ipynb
+   tutorials/timeseries.ipynb
+   tutorials/grid_data.ipynb
+   tutorials/cactus_grid_functions.ipynb
+   tutorials/cactus_horizons.ipynb
+   tutorials/gravitational_waves.ipynb
 
 Reference material (classes, functions, ...)
 ---------------------------------------------
