+++ _sources/tutorials/timeseries.ipynb.txt
@@ -1,0 +1,422 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Working with time series, frequency series, and unit conversion"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "In this notebook, we show some of the most useful features of the [timeseries](../series.rst) module. To do so, we will analyze a fake gravitational-wave signal. We will also show the [frequencyseries](../series.rst) module and the [unitconv](../unitconv.rst) modules.\n",
+    "\n",
+    "First, let's generate this signal. \n",
+    "\n",
+    "(This notebook is meant to be converted in Sphinx documentation and not used directly.)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import matplotlib.pyplot as plt\n",
+    "import numpy as np\n",
+    "from kuibit import timeseries as ts\n",
+    "from kuibit import series\n",
+    "from kuibit import unitconv as uc\n",
+    "from kuibit.gw_utils import luminosity_distance_to_redshift\n",
+    "\n",
+    "%matplotlib inline"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "t = np.linspace(0, 20, 5000)\n",
+    "y = np.sin(t)\n",
+    "\n",
+    "# Generate a TimeSeries by providing the times and the values of the series\n",
+    "gw = ts.TimeSeries(t, y)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "To access the times and the values, use ``gw.t`` and ``gw.y``. "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "def plot(ser, lab1=\"d h\", lab2=\"t\", *args, **kwargs):\n",
+    "    \"\"\"Plot Series ser with labels\"\"\"\n",
+    "    plt.ylabel(lab1)\n",
+    "    plt.xlabel(lab2)\n",
+    "    plt.plot(ser, *args, **kwargs)\n",
+    "    \n",
+    "plot(gw)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Let's make this more interesting: we square the signal and add Gaussian noise. [TimeSeries](../timeseries_ref.html#kuibit.timeseries.TimeSeries) support directly using all the basic mathematical operations in Python and NumPy, so we can easily square the signal and add another TimeSeries to it."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "noise = np.random.normal(0, 0.1, len(gw))  # Here we used len() on the TimeSeries\n",
+    "noise_ts = ts.TimeSeries(t, noise)\n",
+    "\n",
+    "gw += noise_ts\n",
+    "gw **= 2\n",
+    "\n",
+    "plot(gw)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
[…]
