+++ NEWS.md
@@ -6,6 +6,7 @@
 - Improvements to documentation and docstrings
 #### Features
 - Added `nanmax`, `nanmix`, `abs_nanmax`, `abs_nanmin` to `BaseNumerical`
+- Added support for HDF5 grid arrays
 #### Bug fixes
 - `cactus_horizons` and `cactus_multiploes` now remove duplicate iterations
 #### Breaking changes
