### 2021-02-23 295960d
Pin jedi to 0.17.2
Version 0.18 is incompatible with ipython. This was fixed in
ipython 7.2.0, which is incompatible with python3.6.
