+++ pyproject.toml
@@ -30,6 +30,7 @@ nbsphinx = "^0.7.1"
 jupyter = "^1.0.0"
 matplotlib = "^3.3.2"
 mayavi = "^4.7.2"
+jedi = "^0.17.2"
 
 [tool.poetry.extras]
 full = ["numba", "lalsuite", "pycbc"]
