+++ docs/utils_ref.rst
@@ -1,0 +1,5 @@
+Reference on kuibit.utils
+==================================
+
+.. automodule:: kuibit.utils
+   :members:
