+++ _sources/index.rst.txt
@@ -96,6 +96,7 @@ Usage
    unitconv.rst
    visualize_matplotlib.rst
    argparse_helper.rst
+   utils.rst
 
 Tutorials
 ---------
@@ -137,6 +138,7 @@ Reference material (classes, functions, ...)
    unitconv_ref.rst
    visualize_matplotlib_ref.rst
    argparse_helper_ref.rst
+   utils_ref.rst
 
 What is a kuibit?
 -----------------
