### 2021-03-03 76e0287
Add entry in TODO to transparently handle slicing of available data
