### 2021-03-04 6bc4955
Add support for slicing HierarchicalGridData
