### 2021-03-04 2fcd77e
Don't set aspect_ratio if it is None
