### 2021-03-04 7088781
Add Python slicing of grid data to TODO
