+++ README.md
@@ -107,5 +107,6 @@ tool that you can use to collect the fruit of your `Cactus` simulations.
 `kuibit` follows the same design and part of the implementation details of
 `PostCactus`, code developed by Wolfgang Kastaun. This fork completely rewrites
 the original code, adding emphasis on documentation, testing, and extensibility.
-The logo contains elements designed by [freepik.com](freepik.com).
-
+The logo contains elements designed by [freepik.com](freepik.com). We thank
+``kuibit`` first users, Stamatis Vretinaris and Pedro Espino, for providing
+comments to improve the code and the documentation.
