### 2021-03-05 b1c049f
Add shape_time_at_iteration
