### 2021-03-08 00b3727
Fix add_text_to_figure_corner for 3D plots
