+++ kuibit/visualize_matplotlib.py
@@ -26,6 +26,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import tikzplotlib
 from mpl_toolkits.axes_grid1 import make_axes_locatable
+from mpl_toolkits.mplot3d import Axes3D
 
 from postcactus import grid_data as gd
 from postcactus.cactus_grid_functions import BaseOneGridFunction
@@ -367,7 +368,9 @@ def plot_colorbar(mpl_artist, figure=None, axis=None, label=None, **kwargs):
 def add_text_to_figure_corner(text, figure=None, axis=None):
     """Add text to the bottom right corner of a figure."""
 
-    return axis.text(
+    text_function = axis.text2D if isinstance(axis, Axes3D) else axis.text
+
+    return text_function(
         0.98,
         0.02,
         text,
