### 2021-03-08 87df201
Add shape_outline_at_time
