+++ NEWS.md
@@ -9,6 +9,8 @@
 - Added support for HDF5 grid arrays
 - Added slicing `HierarchicalGridData`
 - Add `shape_time_at_iteration` in `cactus_horizon`
+- Add `shape_at_time` in `cactus_horizon`
+- Add `shape_outline_at_time` in `cactus_horizon`
 #### Bug fixes
 - `cactus_horizons` and `cactus_multiploes` now remove duplicate iterations
 #### Breaking changes
