### 2021-03-08 fe51198
Avoid unprotected next (even if it should not give problems)
