### 2021-03-09 a42fa20
Add first rough implementation of plot_apparent_horizon
