+++ kuibit/visualize_mayavi.py
@@ -40,6 +40,78 @@ def disable_interactive_window():
     """
     mlab.options.offscreen = True
 
+def plot_apparent_horizon(horizon, iteration, color=None, **kwargs):
+    """Plot given apparent horizon.
+
+    TODO: This is very tentative and will change in the future!
+
+    Unknown arguments are passed to ``mayavi.mesh``. One of these arguments
+    can be ``color``. You can set ``color`` to
+
+    :param horizon: Apparent horizon to plot.
+    :type horizon: `~.:py:class:OneHorizon`
+    :param color: RGB values of the color of the horizon.
+                  ``(1,1,1)`` for a white hole, ``(0,0,0)`` for a black hole.
+    :type color: tuple of three floats
+    :param iteration: Iteration to plot.
+    :type iteration: int
+    """
+    if not isinstance(horizon, OneHorizon):
+        raise TypeError("Invalid horizon")
+
+    # We add this here because one could set colormap instead of color.
+    if color is None:
+        color = (0, 0, 0)
+
+    # TODO (FEATURE): Add time as parameter alternative to iteration.
+
+    # HACK: We are plotting multiple patches on top of each other.
+    #
+    # Apparent horizons are described by multiple patches. At the moment, kuibit
+    # cannot merge them in a meaningful way. So, we cheat and plot everything
+    # with the same color so that it is impossible to see that there are
+    # multiple patches.
+
+    # First, we get the patches
+    #
+    # patches is a list of three elements, the three coordinates. Each coordinate
+    # is a list with the various patches describes as a list of np.array.
+    patches = horizon.shape_at_iteration(iteration)
+
+    # We must remove one layer from patches, the one that describes the various
+    # patches. We do this in a rather brutal way by looping over all the
+    # elements.
+    #
+    # What is going on here? We have an other loop (dim in range(3)) that loops
+    # over the three coordinates. The elements of this list comprehension are
+    # np.concatenate of the various points in the patch.
+    shape_xyz = [
+        np.concatenate(patch for patch in patches[dim]) for dim in range(3)
+    ]
+
+    # Here we plot!
+    return mlab.mesh(*shape_xyz, color=color, **kwargs)
+
+
+def add_text_to_figure_corner(text, color=(1, 1, 1), **kwargs):
+    """Add text to the bottom right corner of a figure.
+
+    TODO: This is very tentative and will change in the future!
+
+    Unknown arguments are passed to mayavi.text.
+
+    :param text: Text to be inserted.
+    :type text: str
+    :param color: RGB colors of the text ``(1,1,1)`` for a white,
+                  ``(0,0,0)`` for black.
+    :type color: tuple of three floats
+
+    :returns: mayavi's Text object
+    :rtype: mayavi.Text
+    ``"""
+
+    return mlab.text(0.0, 0.93, text, color=color, **kwargs)
+
 def save(outputpath, figure_extension):
     """Save figure to outputpath with figure_extension.
 
