### 2021-03-10 1bea74c
Add compute_horizons_separation
