+++ tests/test_cactus_horizons.py
@@ -417,3 +417,13 @@ class TestOneHorizon(unittest.TestCase):
                 self.ho.shape_outline_at_iteration(1024, [None, None, 0]),
             )
         )
+
+    def test_compute_horizons_separation(self):
+        ho1 = self.hor[(0, 1)]
+        ho2 = self.hor[(0, 2)]
+
+        separation = ch.compute_horizons_separation(ho1, ho2)
+
+        self.assertEqual(len(separation), 552)
+        # They are the same file, so the separation should be 0 everywhere
+        self.assertAlmostEqual(separation.max(), 0)
