### 2021-03-14 Sbozzolo/kuibit#12 @Yurlungur
[JOSS-Paper] comparison to the state of the field
Overall the paper reads quite well. One thing I noticed is that there is not much comparison to other visualization and analysis tools. For example, how does kuibit compare to visit, yt, or even pycactus? It may also be enough to mention that many users of the Einstein Toolkit write their own visualization pipelines by hand.
