### 2021-03-14 251ed55
Allow save in visualize_mayavi to take arbitrary arguments
