### 2021-03-14 1ec0d00
Add paragraph on alternatives
Fixes #12
