+++ paper.bib
@@ -71,3 +71,55 @@
   year =         2005,
   eprint =       "arXiv:astro-ph/0503420",
 }
+
+@misc{rugutils,
+  author = {Guercilena, Federico},
+  title = {PyCactus},
+  publisher = {BitBucket},
+  journal = {BitBucket repository},
+  url = {https://bitbucket.org/relastro/rugutils}
+}
+
+@misc{scidata,
+  author = {Radice, David},
+  title = {scidata},
+  publisher = {BitBucket},
+  journal = {BitBucket repository},
+  url = {https://bitbucket.org/dradice/scidata}
+}
+
+@misc{pyGWAnalysis,
+  author = {Reisswig, Christian},
+  title = {pyGWAnalysis},
+  publisher = {Einstein Toolkit},
+  journal = {SVN repository},
+  url = {https://svn.einsteintoolkit.org/pyGWAnalysis/}
+}
+
+@InCollection{visit,
+author = {Hank Childs and Eric Brugger and Brad Whitlock and Jeremy Meredith and Sean Ahern and David Pugmire and Kathleen Biagas and Mark Miller and Cyrus Harrison and Gunther H. Weber and Hari Krishnan and Thomas Fogal and Allen Sanderson and Christoph Garth and E. Wes Bethel and David Camp and Oliver R\"{u}bel and Marc Durant and Jean M. Favre and Paul Navr\'{a}til},
+title = {{VisIt: An End-User Tool For Visualizing and Analyzing Very Large Data}},
+year = "2012",
+pages = "357-372",
+month = "Oct",
+booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific Insight}},
+}
+
+@ARTICLE{POWER,
+       author = {{Johnson}, Daniel and {Huerta}, E.~A. and {Haas}, Roland},
+        title = "{Python Open source Waveform ExtractoR (POWER): an open source, Python package to monitor and post-process numerical relativity simulations}",
+      journal = {Classical and Quantum Gravity},
+     keywords = {General Relativity and Quantum Cosmology, Computer Science - Computational Engineering, Finance, and Science, Physics - Computational Physics, C.0, G.1.0, J.2},
+         year = 2018,
+        month = jan,
+       volume = {35},
+       number = {2},
+          eid = {027002},
+        pages = {027002},
+          doi = {10.1088/1361-6382/aa9cad},
+archivePrefix = {arXiv},
+       eprint = {1708.02941},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018CQGra..35b7002J},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
