### 2021-03-14 Sbozzolo/kuibit#12 @Sbozzolo
Thanks, I added a paragraph surveying the alternatives.

You can find the diff here: https://github.com/Sbozzolo/kuibit/compare/joss-paper?expand=1
