### 2021-03-14 #3099 @Sbozzolo
Thanks, I added a paragraph that lists all the software officially recommended in this page https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing.

The only other software that is comparable to `kuibit` is `SimulationTools`, but it requires Mathematica (non-free). Most of the other software for analysis appears to be outdated, unmaintained, and/or lacking documentation (including `PyCactus`). Tools like `visit` cannot be used to do quantitative analysis but only visualization. So, I think that `kuibit` is unique, especially in its mission of being a package accessible to newcomers (and thus, its emphasis on documentation, tutorials, example, comments in the code, and in requiring no proprietary software).
