### 2021-03-14 Sbozzolo/kuibit#13 @Yurlungur
[JOSS-paper] questions about contributor guidelines?
Your contributing document already contains lots of great info. But there are a couple of things that I think might be missing. What is your procedure for merge requests? Do tests have to pass? How many people have to review a piece of code? Do you prefer merge requests from forks or branches? etc.
