### 2021-03-14 #3099 @Yurlungur
Your contribution rules also look quite good. I added an issue with one question I had about them [here](https://github.com/Sbozzolo/kuibit/issues/13).
