### 2021-03-15 f63db8f
Update CONTRIBUTING
Fixes #13
