### 2021-03-15 #3099 @Sbozzolo
Thanks, I updated the `CONTRIBUTING` document with a detailed list of the steps to contribute.
