### 2021-03-15 ce1e566
Add ignore_symlinks to SimDir
