+++ _sources/index.rst.txt
@@ -43,19 +43,11 @@ For a full list of available features, see the :doc:`features page <features>`.
   ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Handle unit conversion, in particular from geometrized to physical
   (:py:mod:`~.unitconv`).
-- [EXPERIMENTAL] Visualize data with ``matplotlib``
-  (:py:mod:`~.visualize_matplotlib`).
-- [EXPERIMENTAL] Write command-line scripts (:py:mod:`~.argparse_helper`).
-- [EXPERIMENTAL] Make movies with  `motionpicture`_.
-- [EXPERIMENTAL] Run full analyses with  `ciak`_.
-
-.. _motionpicture: https://github.com/Sbozzolo/motionpicture
-.. _ciak: https://github.com/Sbozzolo/ciak
 
 Installation
 ------------
 
-``kuibit`` is available in PyPI. To install it with ``pip``
+``kuibit`` is available in PyPI. To install it with `pip`
 
 .. code-block:: bash
 
@@ -96,9 +88,6 @@ Usage
    grid_data.rst
    sensitivity_curves.rst
    unitconv.rst
-   visualize_matplotlib.rst
-   argparse_helper.rst
-   utils.rst
 
 Tutorials
 ---------
@@ -112,9 +101,6 @@ Tutorials
    tutorials/cactus_grid_functions.ipynb
    tutorials/cactus_horizons.ipynb
    tutorials/gravitational_waves.ipynb
-
-You can find some real-word examples in the folder ``examples`` in the
-``experimental`` branch.
 
 Reference material (classes, functions, ...)
 ---------------------------------------------
@@ -138,9 +124,6 @@ Reference material (classes, functions, ...)
    grid_data_ref.rst
    sensitivity_curves_ref.rst
    unitconv_ref.rst
-   visualize_matplotlib_ref.rst
-   argparse_helper_ref.rst
-   utils_ref.rst
 
 What is a kuibit?
 -----------------
