+++ _sources/features.rst.txt
@@ -149,3 +149,20 @@ Horizons
 - Read and represent the ASCII output from ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Work with the shape of the horizons and their properties (as timeseries).
 - Cut the 3D shape into 2D projection along the axes centered in the origin of the horizon.
+
+Experimental
+------------
+
+Experimental features may change in the future.
+
+- Visualize grid data.
+- Build command-line scripts with common options.
+- Make movies with `motionpicture`_.
+
+.. _motionpicture: https://github.com/Sbozzolo/motionpicture
+
+- Run full analyses with `ciak`_.
+
+.. _ciak: https://github.com/Sbozzolo/ciak
+
+- Produce log messages.
