### 2021-03-16 cc74dbb
Fix typo in docstring in cactus_waves
