### 2021-04-02 19cb045
Add plot_horizon_shape
