+++ _sources/testimonials.rst.txt
@@ -7,7 +7,7 @@ email or Telegram.
 
     "I wish I had kuibit when I started working with the Einstein Toolkit."
 
-    -- Gabriele Bozzola, PhD candidate at The University of Arizona (2021)
+    -- Gabriele Bozzola, PhD candidate at The University of Arizona (2020)
 
     "As a PhD with no previous experience in numerical relativity, kuibit enabled
     me to analyze simulations, instead of spending months to understand and
