### 2021-04-02 103e114
Add plot_ah_trajectory
