### 2021-04-06 dbe5c1b
Add create_figure
