### 2021-04-06 #3099 @eloisabentivegna
@Sbozzolo, I am almost done with the code review.

I have a few suggestions regarding the paper text:
- The Einstein Toolkit should always have a determinate article, e.g. "made by Einstein Toolkit" -> "made by the Einstein Toolkit" on line 28.
- On line 55, "visit" -> "VisIt"
- The citation label for Frontera should be corrected
- The description of a typical workflow seems to imply that Simulation Factory is required to use kuibit. Is this the case? If so, it should be stated clearly.

I am also not sure I understand the note

`(This notebook is meant to be converted in Sphinx documentation and not used directly.)`

in the tutorial notebooks. Are these really not supposed to be used as tutorials? If not, what other document can users refer to for real-world examples?
