### 2021-04-06 #3099 @Sbozzolo
Thanks, I took care of your comments. I added a sentence to say that `kuibit` does not assume any specific simulation workflow but gathers information from filenames and metadata.

> I am also not sure I understand the note
>
> (This notebook is meant to be converted in Sphinx documentation and not used directly.)
>
> in the tutorial notebooks. Are these really not supposed to be used as tutorials?

The documentation notebooks are generated as part of the continuous integration pipeline using data that I include in the repo. They serve the purpose of showing the basic features available and how they should be used, but, due to technical limitations, the data included is very minimal. There is little point in running the notebooks interactively, as some features will fail, given that the data is not realistic.

One can find real-world code looking at the `experimental` branch, which contains a large number of [examples](https://github.com/Sbozzolo/kuibit/tree/experimental/examples). The reason these are in the `experimental` branch is because all these scripts rely on at least one additional modules, `argparse_helper`, which I want to extend further before merging it back to `master`. This module is used to work with command-line arguments, and does not affect much the rest of the code, so the files are very good examples to learn how to work with `kuibit`. They are as real-world as it gets: if one is using the `experimental` branch, they can be used directly to do science. One can find this piece of information in the [documentation](https://sbozzolo.github.io/kuibit/#tutorials). The readme of the repo also tries to make users aware of the `experimental` branch.

Since you couldn't find the examples, I guess I should make it more explicit.
