### 2021-04-07 4b36cad
Add plot_horizon_vector
