### 2021-04-08 #3099 @eloisabentivegna
> Thanks, I took care of your comments. I added a sentence to say that `kuibit` does not assume any specific simulation workflow but gathers information from filenames and metadata.

Thanks!

> Since you couldn't find the examples, I guess I should make it more explicit.

I agree. It's nice that there are pointers to the experimental branch and that users are told there are real-world examples there, but this is easy to miss at the moment. I suggest changing the section header in the readme from "Experimental Branch" to "Tutorials" (or something like that), and perhaps adding to the note in the docs/tutorials something like "For real-world tutorials, please see ...".

What do you think?
