+++ docs/index.rst
@@ -74,6 +74,13 @@ place where to discuss it. Alternatively, you can also open an issue on GitHub.
 Frequently asked questions are collected in the page :ref:`faq:Frequently Asked
 Questions`.
 
+In addition to the tutorials that are presented in this page, real-world
+examples can be found `examples
+<https://github.com/Sbozzolo/kuibit/tree/experimental/examples>`_ folder of the
+``experimental`` branch. See the section `Experimental branch and examples
+<https://github.com/Sbozzolo/kuibit/#experimental-branch-and-examples>`_ for
+more information.
+
 Usage
 -----
 
@@ -105,12 +112,21 @@ Tutorials
    tutorials/cactus_horizons.ipynb
    tutorials/gravitational_waves.ipynb
 
+In addition to these tutorials, you can find real world examples in the
+`examples <https://github.com/Sbozzolo/kuibit/tree/experimental/examples>`_
+folder of the ``experimental`` branch. See the section `Experimental branch and
+examples
+<https://github.com/Sbozzolo/kuibit/#experimental-branch-and-examples>`_ for
+more information.
+
+
 Reference material (classes, functions, ...)
 ---------------------------------------------
 
 .. toctree::
    :maxdepth: 1
 
+   testimonials.rst
    features.rst
    faq.rst
    simdir_ref.rst
