+++ _sources/index.rst.txt
@@ -46,19 +46,11 @@ For a full list of available features, see the :doc:`features page <features>`.
   ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Handle unit conversion, in particular from geometrized to physical
   (:py:mod:`~.unitconv`).
-- [EXPERIMENTAL] Visualize data with ``matplotlib``
-  (:py:mod:`~.visualize_matplotlib`).
-- [EXPERIMENTAL] Write command-line scripts (:py:mod:`~.argparse_helper`).
-- [EXPERIMENTAL] Make movies with  `motionpicture`_.
-- [EXPERIMENTAL] Run full analyses with  `ciak`_.
-
-.. _motionpicture: https://github.com/Sbozzolo/motionpicture
-.. _ciak: https://github.com/Sbozzolo/ciak
 
 Installation
 ------------
 
-``kuibit`` is available in PyPI. To install it with ``pip``
+``kuibit`` is available in PyPI. To install it with `pip`
 
 .. code-block:: bash
 
@@ -82,6 +74,13 @@ place where to discuss it. Alternatively, you can also open an issue on GitHub.
 Frequently asked questions are collected in the page :ref:`faq:Frequently Asked
 Questions`.
 
+In addition to the tutorials that are presented in this page, real-world
+examples can be found `examples
+<https://github.com/Sbozzolo/kuibit/tree/experimental/examples>`_ folder of the
+``experimental`` branch. See the section `Experimental branch and examples
+<https://github.com/Sbozzolo/kuibit/#experimental-branch-and-examples>`_ for
+more information.
+
 Usage
 -----
 
@@ -99,9 +98,6 @@ Usage
    grid_data.rst
    sensitivity_curves.rst
    unitconv.rst
-   visualize_matplotlib.rst
-   argparse_helper.rst
-   utils.rst
 
 Tutorials
 ---------
@@ -116,8 +112,13 @@ Tutorials
    tutorials/cactus_horizons.ipynb
    tutorials/gravitational_waves.ipynb
 
-You can find some real-word examples in the folder ``examples`` in the
-``experimental`` branch.
+In addition to these tutorials, you can find real world examples in the
+`examples <https://github.com/Sbozzolo/kuibit/tree/experimental/examples>`_
+folder of the ``experimental`` branch. See the section `Experimental branch and
+examples
+<https://github.com/Sbozzolo/kuibit/#experimental-branch-and-examples>`_ for
+more information.
+
 
 Reference material (classes, functions, ...)
 ---------------------------------------------
@@ -125,6 +126,7 @@ Reference material (classes, functions, ...)
 .. toctree::
    :maxdepth: 1
 
+   testimonials.rst
    features.rst
    faq.rst
    simdir_ref.rst
@@ -141,9 +143,6 @@ Reference material (classes, functions, ...)
    grid_data_ref.rst
    sensitivity_curves_ref.rst
    unitconv_ref.rst
-   visualize_matplotlib_ref.rst
-   argparse_helper_ref.rst
-   utils_ref.rst
 
 What is a kuibit?
 -----------------
