+++ _sources/features.rst.txt
@@ -92,14 +92,13 @@ Multipoles and waves
   - strains at given multipole number,
   - strains at a given point (accounting for the spin-weigthed spherical harmonics),
   - strain as observed by LIGO/Virgo (considering the antenna patterns),
-  - force/linear momentum along the z axis lost via gravitational waves (one or multiple modes),
   - power/energy lost via gravitational waves (one or multiple modes),
   - torque/angular momentum along the z axis lost via gravitational waves (one or multiple modes),
   - Compute the last two for electromagnetic waves from ``Phi2``.
 
 - Extrapolate waves at infinity with polynomial expansion in real/imaginary parts or amplitude and phase.
 - Compute spin-weigthed spherical harmonics.
-- Convert from RA and Dec to spherical coordinates.
+- Convert from RA and Dec to spherical coordiantes.
 - Compute antenna responses.
 - Compute signal to noise ratio from strain.
 - Compute redshift from luminosity distance.
@@ -150,20 +149,3 @@ Horizons
 - Read and represent the ASCII output from ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Work with the shape of the horizons and their properties (as timeseries).
 - Cut the 3D shape into 2D projection along the axes centered in the origin of the horizon.
-
-Experimental
-------------
-
-Experimental features may change in the future.
-
-- Visualize grid data.
-- Build command-line scripts with common options.
-- Make movies with `motionpicture`_.
-
-.. _motionpicture: https://github.com/Sbozzolo/motionpicture
-
-- Run full analyses with `ciak`_.
-
-.. _ciak: https://github.com/Sbozzolo/ciak
-
-- Produce log messages.
