### 2021-04-08 #3099 @Sbozzolo
> I agree. It's nice that there are pointers to the experimental branch and that users are told there are real-world examples there, but this is easy to miss at the moment. I suggest changing the section header in the readme from "Experimental Branch" to "Tutorials" (or something like that), and perhaps adding to the note in the docs/tutorials something like "For real-world tutorials, please see ...".
>
> What do you think?

I added links to the examples in more places (e.g., the help section, at the bottom of the tutorials, in the opening of the readme), and expanded the section in the readme. (commit https://github.com/Sbozzolo/kuibit/commit/858bdf9e5f0526d72db5ba393ac393607131209b). Hopefully this helps.
