### 2021-04-09 ac0e80c
Fix grammar in a docstring
