### 2021-04-09 2b1df5e
Add dimensions to add_grid_parsers
