### 2021-04-09 Sbozzolo/kuibit#15 @Yurlungur
A question on developing with poetry?
I know poetry isn't yours---it's essentially a dependency. But from skimming Poetry's documentation, and yours, one thing isn't clear to me, which might be worth highlighting in the documentation.

When I develop a python package---e.g., yt---with a traditional `setup.py` build system, I can call
```python
python setup.py develop --user
```
to tell it to create symlinks to the python code so that when I import a module from anywhere, if I modify the code in the repo, the changes are reflected in the import. Is the same true here? i.e., do I have to do anything special in poetry to enable this kind of behaviour?
