### 2021-04-09 #3099 @Yurlungur
@Sbozzolo I think I've finished my review. I found a few very minor things. Mainly, when installing from poetry for development, I got hung up on a couple related details. The installation appears to be a virtualenv in the source directory, so I need to call `poetry shell` from within that directory, and python imports only work from there. I submit a PR to update the `README` [here](https://github.com/Sbozzolo/kuibit/pull/16) and a further question about the `poetry` equivalent of `setup.py develop` [here](https://github.com/Sbozzolo/kuibit/issues/15)

I also had some idle thoughts/questions, which I submitted as issues. They are not blocking and not really related to the review.

Other than that, I'm pretty happy with everything. I managed to use your documentation/tutorials to visualize a couple simple examples I had lying around. A BHNS merger, a TOV star, and the GW150914 gallery example from the ET website. All of it worked painlessly, thanks to the documentation and the inline docstrings. Overall, very impressive.
