### 2021-04-10 b37b09c
Rename _derive methods to _differentiate
Thanks @Yurlungur for pointing this out
