+++ _sources/index.rst.txt
@@ -32,7 +32,7 @@ For a full list of available features, see the :doc:`features page <features>`.
 - Work with the power spectral densities of known detectors
   (:py:mod:`~.sensitivity_curves`).
 - Represent and manipulate time series (:py:mod:`~.timeseries`). Examples of
-  functions available for time series: ``integrate``, ``derive``, ``resample``,
+  functions available for time series: ``integrate``, ``differentiate``, ``resample``,
   ``to_FrequencySeries`` (Fourier transform).
 - Represent and manipulate frequency series (:py:mod:`~.frequencyseries`), like
   Fourier transforms of time series. Inverse Fourier transform is available.
@@ -46,19 +46,11 @@ For a full list of available features, see the :doc:`features page <features>`.
   ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Handle unit conversion, in particular from geometrized to physical
   (:py:mod:`~.unitconv`).
-- [EXPERIMENTAL] Visualize data with ``matplotlib``
-  (:py:mod:`~.visualize_matplotlib`).
-- [EXPERIMENTAL] Write command-line scripts (:py:mod:`~.argparse_helper`).
-- [EXPERIMENTAL] Make movies with  `motionpicture`_.
-- [EXPERIMENTAL] Run full analyses with  `ciak`_.
-
-.. _motionpicture: https://github.com/Sbozzolo/motionpicture
-.. _ciak: https://github.com/Sbozzolo/ciak
 
 Installation
 ------------
 
-``kuibit`` is available in PyPI. To install it with ``pip``
+``kuibit`` is available in PyPI. To install it with `pip`
 
 .. code-block:: bash
 
@@ -106,9 +98,6 @@ Usage
    grid_data.rst
    sensitivity_curves.rst
    unitconv.rst
-   visualize_matplotlib.rst
-   argparse_helper.rst
-   utils.rst
 
 Tutorials
 ---------
@@ -129,6 +118,7 @@ folder of the ``experimental`` branch. See the section `Experimental branch and
 examples
 <https://github.com/Sbozzolo/kuibit/#experimental-branch-and-examples>`_ for
 more information.
+
 
 Reference material (classes, functions, ...)
 ---------------------------------------------
@@ -153,9 +143,6 @@ Reference material (classes, functions, ...)
    grid_data_ref.rst
    sensitivity_curves_ref.rst
    unitconv_ref.rst
-   visualize_matplotlib_ref.rst
-   argparse_helper_ref.rst
-   utils_ref.rst
 
 What is a kuibit?
 -----------------
