+++ _sources/tutorials/timeseries.ipynb.txt
@@ -164,8 +164,8 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "gw_spline_der = gw.spline_derived()\n",
-    "gw_numer_der = gw.derived()\n",
+    "gw_spline_der = gw.spline_differentiated()\n",
+    "gw_numer_der = gw.differentiated()\n",
     "\n",
     "plot(gw_spline_der, label='spline')\n",
     "plot(gw_numer_der, lab1=\"derivative (dh)\", label='finite')\n",
