### 2021-04-10 Sbozzolo/kuibit#15 @Sbozzolo
This is a long-standing question/issue with poetry (https://github.com/python-poetry/poetry/issues/34). I think that now it is possible to install a package in editable mode (once I managed to get it to work), but I cannot tell you exactly how to do it.

`poetry` installs a virtualenv in `~/.cache/pypoetry/*/virtualenvs` (at least on GNU/Linux). My personal development workflow is to activate that environment directly or to use `poetry shell`. Note that once you are in the `poetry shell` you can freely move around your filesystem, the experience is quite similar to installing a package in development mode.
