### 2021-04-10 #3099 @Sbozzolo
Thank you very much for your comments, @Yurlungur!

As for `poetry`. When you are in a `poetry shell`, the imports should work anywhere.  When you are not in a `poetry shell`, the imports will work only in the source directory.

I realize that people might not be familiar with `poetry`, and I will definitely add more detailed instruction if I see that this is a point of struggle for (prospective) contributors.
