### 2021-04-10 #3099 @Yurlungur
Hmm. I think I tried importing kuibit from within the poetry shell but in a different directory and it didn't work. But maybe I'm mistaken.
