### 2021-04-10 #3099 @Sbozzolo
It is true that you have to run `poetry shell` from the directory where the code lives, but, once you have done it, you should be able to access `kuibit` from anywhere. `poetry shell` essentially activates a virtualenv.
