### 2021-04-10 Sbozzolo/kuibit#15 @Yurlungur
Ah ok. Cool. That works, then. It might be nice to add a comment about activating the environment directly in your documentation.
