### 2021-04-10 Sbozzolo/kuibit#20 @eloisabentivegna
Two failing tests.
Tests `test_antenna_responses` and `test_ra_dec_to_theta_phi` are failing for me at the moment, with errors:

AssertionError: 0.1676658308656924 != 0.173418558 within 7 places (0.005752727134307606 difference)

and

AssertionError: 2.458839344228527 != 2.3697174 within 7 places (0.08912194422852693 difference)

respectively.

Could this be the effect of any of the changes implemented as a result of the review in https://github.com/openjournals/joss-reviews/issues/3099?
