### 2021-04-10 Sbozzolo/kuibit#20 @Sbozzolo
That's quite interesting. All the tests are running fine on my computers and the supercomputers I have access to, and on GitHub actions.

![2021-04-10:08:41:21](https://user-images.githubusercontent.com/9167485/114275786-ad098500-99d8-11eb-8c5a-9473325b01fa.png)

The failure of the second test leads to the failure of the first. So, I have to understand why the second fails. The only possible reason I can think is that the `datetime` Python module is behaving differently on your system. Other than that, the function only does numerical manipulations of numbers with `numpy` and it is very difficult to see how they could lead to different results.

Can you add the following print statement after line 290 of `gw_utils.py` and run the test again?
```python
print(base_posix_date, posix_date)
```
On my computer, it prints `946753200.0 1442249445.0`.

Are you running the test in the poetry virtual environment? What version of Python and OS are you using?
