### 2021-04-11 Sbozzolo/kuibit#20 @eloisabentivegna
Sure thing. I get

> .............................................................946728000.0 1442220645.0
> 946728000.0 1442220645.0
> ..........................................................................946728000.0 1442220645.0
> 946728000.0 1442220645.0
> 946728000.0 1442220645.0
> 946728000.0 1442220645.0
> .946728000.0 1442220645.0
> 946728000.0 1442220645.0
> 946728000.0 1442220645.0
> 946728000.0 1442220645.0
> ...946728000.0 1442220645.0
> F..946728000.0 1442220645.0
> F..............................................................
> ======================================================================
> FAIL: test_antenna_responses (tests.test_gw_utils.TestGWUtils)
> ----------------------------------------------------------------------

I am using Python 3.8 on OS 10.15.
