### 2021-04-11 Sbozzolo/kuibit#20 @Sbozzolo
Yes, they are different! This might be related to some behavior specific to macOS and how it defines local times.

Can you please try to change the two lines that define `base_posix_data` and `posix_date` to
```python
    base_posix_date = datetime.datetime.strptime(
        "2000-01-01 12:00:00 UTC", "%Y-%m-%d %H:%M:%S %Z"
    ).timestamp()
    posix_date = datetime.datetime.strptime(
        time_utc + " UTC", "%Y-%m-%d %H:%M:%S %Z"
    ).timestamp()
```
I am forcing the timezone to be UTC. Hopefully this works.
