### 2021-04-11 210012c
Add macOS to CI
