+++ .github/workflows/test.yml
@@ -5,7 +5,10 @@ on: [push]
 jobs:
 
   Test:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
 
     steps:
       - uses: actions/checkout@v2
@@ -46,36 +49,9 @@ jobs:
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
+        if: runner.os == 'Linux'
         with:
           file: coverage.xml
           flags: unittests
           env_vars: OS,PYTHON
           fail_ci_if_error: true
-
-      - name: Produce documentation
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y pandoc
-          sudo apt-get install -y xvfb
-          cd docs/
-          xvfb-run poetry run make html
-          cd ..
-
-      - name: Commit documentation changes
-        run: |
-          git clone https://github.com/sbozzolo/PostCactus.git --branch gh-pages --single-branch gh-pages
-          cp -r docs/_build/html/* gh-pages/
-          cd gh-pages
-          git config --local user.email "action@github.com"
-          git config --local user.name "GitHub Action"
-          git add .
-          git commit -m "Update documentation" -a || true
-          # The above command will fail if no changes were present, so we ignore
-          # the return code.
-      - name: Push changes
-        uses: ad-m/github-push-action@master
-        with:
-          branch: gh-pages
-          # force: true
-          directory: gh-pages
-          github_token: ${{ secrets.GITHUB_TOKEN }}
