### 2021-04-11 Sbozzolo/kuibit#20 @Sbozzolo
I added macOS to the CI and the tests are all passing even without that change. There must be a subtle problem, but I hope that the proposed change will fix it.
