### 2021-04-11 Sbozzolo/kuibit#20 @eloisabentivegna
The change does not make a difference (which is at least consistent with your second message). There's probably a version mismatch in some of the dependencies. Given the magnitude of the difference, I am not considering this a blocker for the JOSS review -- test maintenance is a continuing process for evolving software anyway.

Regardless of the review, I am happy to send more info on the other components of my setup, if this can help. Or you can tell me which system configuration you used on macOS and I can figure out where the differences are.
