### 2021-04-11 #3099 @eloisabentivegna
I am through with my review as well. The package is a very nice submission and, judging from the number of people I've heard over the years longing for a comprehensive viz package in Python for the ET, it fills a big gap. Thanks @Sbozzolo for the submission! I recommend acceptance.

The only reservation I have, perhaps along some of the things that @Yurlungur has expressed, is that `poetry` is a bit of a pain point. This would be OK if only people wanting to develop the package had the necessity to use it (although that means adding a barrier to contributions), but at the moment any user wanting to access the examples has to learn to use `poetry`. My recommendation is to merge the experimental features required by the examples into the main branch as soon as possible, so that standard Python installation routes become available.
