### 2021-04-11 #3099 @eloisabentivegna
(Changing hat.)

@Sbozzolo: congratulation on the accepted submission! 🎉

Can you provide a tagged release and code archive details (see https://joss.readthedocs.io/en/latest/submitting.html#the-review-process)?
