### 2021-04-11 #3099 @Sbozzolo
Thanks, @eloisabentivegna. I'll address your comment about `poetry`, I'll fix the DOI, and I'll follow your instruction in the next two hours or so.
