+++ paper.bib
@@ -41,6 +41,7 @@
   year =         2003,
   address =      {Berlin},
   publisher =    {Springer},
+  doi = "10.1007/3-540-36569-9_13"
 }
 
 @article{einsteintoolkit2,
