+++ tests/test_kuibit.py
@@ -22,4 +22,4 @@ from kuibit import __version__
 
 class Testkuibit(unittest.TestCase):
     def test_version(self):
-        self.assertEqual(__version__, "1.0.0b0")
+        self.assertEqual(__version__, "1.0.0")
