+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "kuibit"
-version = "1.0.0b0"
+version = "1.0.0"
 description = "Read and analyze Einstein Toolkit simulations."
 authors = ["Gabriele Bozzola <gabrielebozzola@arizona.edu>"]
 maintainers = ["Gabriele Bozzola <gabrielebozzola@arizona.edu>"]
