### 2021-04-12 #3099 @Sbozzolo
> The only reservation I have, perhaps along some of the things that @Yurlungur has expressed, is that poetry is a bit of a pain point. This would be OK if only people wanting to develop the package had the necessity to use it (although that means adding a barrier to contributions), but at the moment any user wanting to access the examples has to learn to use poetry. My recommendation is to merge the experimental features required by the examples into the main branch as soon as possible, so that standard Python installation routes become available.

My intent is definitely to require `poetry` only for developers. I understand that this may be a barrier, but I think that `poetry` is getting a lot of traction in the Python community (and alternatives equivalent to `poetry` would still require contributors to learn a new tool). Depending on the developer's feedback, I will be happy to provide further instruction on how to use `poetry`. However, given the already considerable amount of information and tutorials available online, my personal point of view is to treat `poetry` like any other tool in the Python ecosystem (e.g., `unittest`).

> but at the moment any user wanting to access the examples has to learn to use poetry.

That's true if you want to run the examples, but you can still read them (and the comments) online.

> so that standard Python installation routes become available.

You can install the `experimental` branch with:
```
pip3 install git+https://github.com/Sbozzolo/kuibit.git@experimental
```

> My recommendation is to merge the experimental features required by the examples into the main branch as soon as possible,

At the moment, the development on the `experimental` branch is driven by my personal needs. I could merge some of the new features back to `master`, but not all the examples could be ported. So, I prefer waiting until things are in a great shape instead of merging work that I am not fully satisfied with. But, I 100% agree that this is a place where things can improved.

> @Sbozzolo: could you also take care of the missing DOI before creating the tagged release and archiving the code?

I fixed that.

> Can you provide a tagged release and code archive details (see https://joss.readthedocs.io/en/latest/submitting.html#the-review-process)?

I did! DOI: 10.5281/zenodo.4679863

Thanks!
