### 2021-04-12 Sbozzolo/kuibit#20 @Sbozzolo
> There's probably a version mismatch in some of the dependencies

If you are running with `poetry`, it should use the very same versions I am using (the ones defined in the `poetry.lock`), so I don't think that's the problem. I would point my finger to some difference between macOS and GNU/Linux, but simply changing the GitHub Action runner to macOS didn't reveal any issue.

>  Or you can tell me which system configuration you used on macOS and I can figure out where the differences are.

I have no idea. I don't have access to any computer with macOS (at least, in an easy way). My test on macOS is with the configurations provided by [this](https://github.com/actions/setup-python), with `poetry`.

Does the program
```python
import datetime
datetime.datetime.strptime( "2000-01-01 12:00:00", "%Y-%m-%d %H:%M:%S").timestamp()
```
return `946753200.0`?
