### 2021-04-12 #3099 @Sbozzolo
Yes, correct.

On Mon, Apr 12, 2021, 00:09 Eloisa Bentivegna ***@***.***>
wrote:

> @Sbozzolo <https://github.com/Sbozzolo>, can I assume 1.0.0 is the tagged
> release you created for this submission?
>
