### 2021-04-12 #3099 @eloisabentivegna
@sbozzolo, one more request: could you change the Zenodo archive title to match this submission title (see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance)? The author list should also match the submission's author list. (But thanks for the credit! :) )
