### 2021-04-12 #3099 @Sbozzolo
Okay, I think I got it (even if I had to regenerate the archive 3 times)!

The new DOI is 10.5281/zenodo.4681119
