+++ _sources/index.rst.txt
@@ -46,11 +46,19 @@ For a full list of available features, see the :doc:`features page <features>`.
   ``QuasiLocalMeasures`` and ``AHFinderDirect``.
 - Handle unit conversion, in particular from geometrized to physical
   (:py:mod:`~.unitconv`).
+- [EXPERIMENTAL] Visualize data with ``matplotlib``
+  (:py:mod:`~.visualize_matplotlib`).
+- [EXPERIMENTAL] Write command-line scripts (:py:mod:`~.argparse_helper`).
+- [EXPERIMENTAL] Make movies with  `motionpicture`_.
+- [EXPERIMENTAL] Run full analyses with  `ciak`_.
+
+.. _motionpicture: https://github.com/Sbozzolo/motionpicture
+.. _ciak: https://github.com/Sbozzolo/ciak
 
 Installation
 ------------
 
-``kuibit`` is available in PyPI. To install it with `pip`
+``kuibit`` is available in PyPI. To install it with ``pip``
 
 .. code-block:: bash
 
@@ -98,6 +106,9 @@ Usage
    grid_data.rst
    sensitivity_curves.rst
    unitconv.rst
+   visualize_matplotlib.rst
+   argparse_helper.rst
+   utils.rst
 
 Tutorials
 ---------
@@ -118,7 +129,6 @@ folder of the ``experimental`` branch. See the section `Experimental branch and
 examples
 <https://github.com/Sbozzolo/kuibit/#experimental-branch-and-examples>`_ for
 more information.
-
 
 Reference material (classes, functions, ...)
 ---------------------------------------------
@@ -143,6 +153,9 @@ Reference material (classes, functions, ...)
    grid_data_ref.rst
    sensitivity_curves_ref.rst
    unitconv_ref.rst
+   visualize_matplotlib_ref.rst
+   argparse_helper_ref.rst
+   utils_ref.rst
 
 What is a kuibit?
 -----------------
