+++ .github/workflows/document.yml
@@ -46,7 +46,7 @@ jobs:
 
       - name: Commit documentation changes
         run: |
-          git clone https://github.com/sbozzolo/PostCactus.git --branch gh-pages --single-branch gh-pages
+          git clone https://github.com/sbozzolo/kuibit.git --branch gh-pages --single-branch gh-pages
           cp -r docs/_build/html/* gh-pages/
           cd gh-pages
           git config --local user.email "action@github.com"
