### 2021-04-12 3956600
Fix get_restarts with 3D files
