+++ paper.bib
@@ -75,7 +75,7 @@
 
 @misc{rugutils,
   author = {Guercilena, Federico},
-  title = {PyCactus},
+  title = {rugutils},
   publisher = {BitBucket},
   journal = {BitBucket repository},
   url = {https://bitbucket.org/relastro/rugutils}
