### 2021-04-13 9c6e18f
Add citation guidelines
