+++ docs/index.rst
@@ -164,3 +164,26 @@ extensibility. The logo contains elements designed by `freepik.com
 <https://freepik.com/>`_. We thank ``kuibit`` first users, Stamatis Vretinaris
 and Pedro Espino, for providing comments to improve the code and the
 documentation.
+
+Citation
+------------
+
+``kuibit`` is built and maintained by the dedication of one graduate student. Please,
+consider citing ``kuibit`` if you find the software useful. You can use the following
+``bibtex`` key.
+
+.. code-block:: bibtex
+
+    @article{kuibit,
+      doi = {10.21105/joss.03099},
+      url = {https://doi.org/10.21105/joss.03099},
+      year = {2021},
+      publisher = {The Open Journal},
+      volume = {6},
+      number = {60},
+      pages = {3099},
+      author = {Gabriele Bozzola},
+      title = {kuibit: Analyzing Einstein Toolkit simulations with Python},
+      journal = {Journal of Open Source Software},
+      version = {1.0.0}
+    }
