+++ README.md
@@ -2,6 +2,7 @@
 <img src="https://github.com/Sbozzolo/kuibit/raw/master/logo.png" height="120">
 </p>
 
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.03099/status.svg)](https://doi.org/10.21105/joss.03099)
 [![codecov](https://codecov.io/gh/Sbozzolo/kuibit/branch/master/graph/badge.svg)](https://codecov.io/gh/Sbozzolo/kuibit)
 ![Tests](https://github.com/Sbozzolo/kuibit/workflows/Tests/badge.svg)
 ![Documentation](https://github.com/Sbozzolo/kuibit/workflows/Document/badge.svg)
@@ -113,7 +114,6 @@ parsing command-line arguments and to produce visualizations, the codes are also
 an excellent material to learn how to use the current stable version of
 `kuibit`.
 
-
 ## What is a _kuibit_?
 
 A kuibit (also known as _kukuipad_, meaning harvest pole) is the tool
@@ -129,3 +129,28 @@ the original code, adding emphasis on documentation, testing, and extensibility.
 The logo contains elements designed by [freepik.com](freepik.com). We thank
 ``kuibit`` first users, Stamatis Vretinaris and Pedro Espino, for providing
 comments to improve the code and the documentation.
+
+## Citation
+
+`kuibit` is built and maintained by the dedication of one graduate student. Please,
+consider citing `kuibit` if you find the software useful. You can use the following
+`bibtex` key.
+``` bibtex
+@article{kuibit,
+  doi = {10.21105/joss.03099},
+  url = {https://doi.org/10.21105/joss.03099},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {60},
+  pages = {3099},
+  author = {Gabriele Bozzola},
+  title = {kuibit: Analyzing Einstein Toolkit simulations with Python},
+  journal = {Journal of Open Source Software},
+  version = {1.0.0},
+  archivePrefix = {arXiv},
+  eprint = {2104.06376},
+  primaryClass = {gr-qc}
+}
+```
+You can find this entry in Python with `from kuibit import __bibtex__`.
