### 2021-04-13 ef62c5f
Add _check_ref_factors
